% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/App.R
\name{App}
\alias{App}
\title{The iNZight App object}
\value{
an object of class 'inzstate', which can be parsed to JSON for piping to other applications (i.e., javascript)
}
\description{
A class representing the current state of the application.
}
\examples{
# the initial state on loading a default application is easy to obtain:
state <- inzight::App()

# alternatively, some things may be specified:
state <- inzight::App(data = iris)

}
\author{
Tom Elliott
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{App$new()}}
\item \href{#method-finalize}{\code{App$finalize()}}
\item \href{#method-clone}{\code{App$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$new(data, con, table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{a data.frame object (or something that can be converted to one)}

\item{\code{con}}{a database connection}

\item{\code{table}}{if using \code{con}, this is the name of the table containing the data; if missing, iNZight will use the first (alphabetically) table in the schema}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-finalize}{}}}
\subsection{Method \code{finalize()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{App$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
